+++ LibreriaRiot/Principal/lobi/Vendedor/CargarCliente.cs
@@ -118,7 +118,7 @@ namespace LibreriaRiot.Principal.lobi.Vendedor
             {
                 msgError("Debe ingresar un teléfono.", lbErrorMenssage6);
             }
-            else if (!int.TryParse(telefonoN, out int telefono))
+            else if (!long.TryParse(telefonoN, out long telefono))
             {
                 msgError("Este campo tiene que ser numérico", lbErrorMenssage6);
             }
